refactor(alarm): sleep 모드 가능 여부
- 이전 코드의 비즈니스적 오류 다수 발생 및 좋지 못한 코드퀄리티로 일괄 변경함
- 최종적으로 66줄 로직은 단 4줄로 처리함

+++ BE/src/main/java/com/sf/honeymorning/alarm/repository/AlarmRepository.java
@@ -23,7 +23,7 @@ public interface AlarmRepository extends JpaRepository<Alarm, Long> {
 		"AND a.wake_up_time = :endTime",
 		nativeQuery = true)
 	List<Alarm> findActiveAlarmsForToday(
-		@Param("dayOfWeekMask") byte dayOfWeekMask,
+		@Param("dayOfWeekMask") Integer dayOfWeekMask,
 		@Param("endTime") LocalTime endTime
 	);
